Move DrawerConfig to external contract
