+++ feature/navigation/drawer/src/main/kotlin/app/k9mail/feature/navigation/drawer/NavigationDrawerExternalContract.kt
@@ -1,8 +1,11 @@
 package app.k9mail.feature.navigation.drawer
 
-import app.k9mail.feature.navigation.drawer.domain.entity.DrawerConfig
-
 interface NavigationDrawerExternalContract {
+
+    data class DrawerConfig(
+        val showUnifiedFolders: Boolean,
+        val showStarredCount: Boolean,
+    )
 
     fun interface DrawerConfigLoader {
         fun loadDrawerConfig(): DrawerConfig
