--- feature/navigation/drawer/src/main/kotlin/app/k9mail/feature/navigation/drawer/domain/entity/DrawerConfig.kt
@@ -1,6 +1,0 @@
-package app.k9mail.feature.navigation.drawer.domain.entity
-
-data class DrawerConfig(
-    val showUnifiedFolders: Boolean,
-    val showStarredCount: Boolean,
-)
