+++ feature/navigation/drawer/src/main/kotlin/app/k9mail/feature/navigation/drawer/domain/usecase/GetDrawerConfig.kt
@@ -1,8 +1,8 @@
 package app.k9mail.feature.navigation.drawer.domain.usecase
 
+import app.k9mail.feature.navigation.drawer.NavigationDrawerExternalContract.DrawerConfig
 import app.k9mail.feature.navigation.drawer.NavigationDrawerExternalContract.DrawerConfigLoader
 import app.k9mail.feature.navigation.drawer.domain.DomainContract.UseCase
-import app.k9mail.feature.navigation.drawer.domain.entity.DrawerConfig
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
