+++ legacy/common/src/main/java/com/fsck/k9/feature/NavigationDrawerConfigLoader.kt
@@ -1,7 +1,7 @@
 package com.fsck.k9.feature
 
+import app.k9mail.feature.navigation.drawer.NavigationDrawerExternalContract.DrawerConfig
 import app.k9mail.feature.navigation.drawer.NavigationDrawerExternalContract.DrawerConfigLoader
-import app.k9mail.feature.navigation.drawer.domain.entity.DrawerConfig
 import com.fsck.k9.K9
 
 class NavigationDrawerConfigLoader : DrawerConfigLoader {
